feat: adding initial update task logic

+++ src/main/java/com/shubham/taskManagement/dao/UserRepo.java
@@ -3,6 +3,6 @@ package com.shubham.taskManagement.dao;
 import com.shubham.taskManagement.models.UserModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepo extends JpaRepository<UserModel, Integer> {
+public interface UserRepo extends JpaRepository<UserModel, Long> {
     UserModel findByEmail(String email);
 }
